pyproject.toml
[project]
name = "visualize"
version = "0.1.0"
description = "Visualize Java program memory state and output a screenshot."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "halo>=0.0.31",
    "install-jdk>=1.1.0",
    "selenium>=4.33.0",
]
